requirements.txt
annotated-types
asgiref
decouple
Django
django-cors-headers
djangorestframework
dj_database_url
greenlet
inflect
psycopg2-binary
pydantic
pydantic_core
python-decouple
pytz
sqlparse
typing_extensions